processing/cloud_function/D002/create-rdf/requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
pymongo==4.8.0
pydantic==2.9.2
google-cloud-run==0.10.14
python-magic==0.4.27
rdflib==7.1.1
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.3
